+++ packages/filepath/src/index.ts
@@ -1,6 +1,8 @@
+import { ParsedPath } from './ParsedPath.js';
+import { ParsedPathError } from './ParsedPathError.js';
 import { allPath, firstPath, resolve } from './parser.js';
-import type { Token } from './Token.js';
-import type { TokenValueType } from './types/TokenValueType.js';
+import type { PathToken } from './Token.js';
+import type { PathTokenValueType } from './types/TokenValueType.js';
 
-export { resolve, firstPath, allPath, type Token };
-export type { TokenValueType };
+export { resolve, firstPath, allPath, ParsedPath, ParsedPathError };
+export type { PathTokenValueType, PathToken };