+++ packages/filepath/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## version "1.0.5" 2025-dec-21
+   ### changed
+   - rename Token to FileToken (Jacob Bogers <jkfbogers@gmail.com>)
+   - export ParsedPath and ParsedPathError (Jacob Bogers <jkfbogers@gmail.com>)
+   - functions "resolve" and "resolveObject" do not return ParsedPathError, instead they throw an error (Jacob Bogers <jkfbogers@gmail.com>)
+
 ## version "1.0.3" 2025-dec-16
    ### changed
    - corrections to README.md document.